tools/mongodb_visual_tool/requirements.txt
pymongo>=4.0.0
Pillow>=9.0.0
python-dotenv>=0.19.0 